Switched to lazy loading for a couple of large libraries
