+++ apps/admin-x-settings/src/components/settings/advanced/CodeInjection.tsx
@@ -2,13 +2,12 @@ import Button from '../../../admin-x-ds/global/Button';
 import CodeEditor from '../../../admin-x-ds/global/form/CodeEditor';
 import CodeModal from './code/CodeModal';
 import NiceModal from '@ebay/nice-modal-react';
-import React, {useRef, useState} from 'react';
+import React, {useMemo, useRef, useState} from 'react';
 import SettingGroup from '../../../admin-x-ds/settings/SettingGroup';
 import TabView from '../../../admin-x-ds/global/TabView';
 import useSettingGroup from '../../../hooks/useSettingGroup';
 import {ReactCodeMirrorRef} from '@uiw/react-codemirror';
 import {getSettingValues} from '../../../api/settings';
-import {html} from '@codemirror/lang-html';
 
 const CodeInjection: React.FC<{ keywords: string[] }> = ({keywords}) => {
     const {
@@ -28,15 +27,17 @@ const CodeInjection: React.FC<{ keywords: string[] }> = ({keywords}) => {
     const headerEditorRef = useRef<ReactCodeMirrorRef>(null);
     const footerEditorRef = useRef<ReactCodeMirrorRef>(null);
 
+    const html = useMemo(() => import('@codemirror/lang-html').then(module => module.html()), []);
+
     const headerProps = {
-        extensions: [html()],
+        extensions: [html],
         hint: 'Code here will be injected into the {{ghost_head}} tag on every page of the site',
         value: headerContent,
         onChange: (value: string) => updateSetting('codeinjection_head', value)
     };
 
     const footerProps = {
-        extensions: [html()],
+        extensions: [html],
         hint: 'Code here will be injected into the {{ghost_foot}} tag on every page of the site',
         value: footerContent,
         onChange: (value: string) => updateSetting('codeinjection_foot', value)
